circle-ci: hide gimp-help-git building step.
I still can't figure out why libxml2 is not importable from Python and
apparently we have reached some concurrency limit in our FLOSS plan on
CircleCI, so I can't use the SSH feature right now to try and debug.
Let's disable building the help for the time being as it is not a
mandatory feature.