Build 2.10.28
- install python 3.9.7
- .circleci: Correctly set $PYTHON variable
- glib-networking is a webkit dependency for libsoup
-.circleci: Comment out webkit

Webkit is failing to build so this comments it out until we have it
building again and against gtk3

- .cricleci: Fix notarize script to use DMG name
- Upgrade create-dmg to version 1.0.9
- .circleci: Never fail signing and notarizing TODO

Other accounts than Infrastructure don't have the right secrets
on circle ci to have this succeed. A better approach would be to
detect that we are on the right account and otherwise avoid
calling the scripts.

However, it wasn't immediately clear how to frame the conditional
so it was skipped.

- .circleci: Use patched mac bundler

  Use version of mac bundler forked by lukaso, while waiting for change
  to be incorporated back to main bundler.

  Fork ensures dependencies in libraries are followed and copied into
  application.

- jhbuildrc-2.10: locked to python2
- Fix bootstrap
- .circleci: Remove test of building gettext
- jhbuildrc: force pygobject and gtk-mac-integration-gtk2-python
  to use python2
- .circleci Modify build script to build 2.10.28
- .circleci: switch to icu from icu55: build failure
- jhbuild: Mismatched gettext version on icu

```
*** Building icu55 *** [1/28]
make CFLAGS="$CFLAGS -DU_CHARSET_IS_UTF8=1 -DU_USING_ICU_NAMESPACE=0"  -j 5
dyld: Library not loaded: /Users/distiller/gtk/inst/lib/libintl.8.dylib
  Referenced from: /Users/distiller/gtk/inst/bin/make
  Reason: Incompatible library version: make requires version 11.0.0 or later, but libintl.8.dylib provides version 10.0.0
```

- .circleci: Remove test of building gettext
- gimp-2-10.bundle: remove libpyglib

Not needed and also not available after build.

- build.sh: Correct path to remove old DMG
- .circleci applescript: allow access to finder.