circleci: update SSL certificates used by Python code.
Though certs seem to be well recognized in curl commands, they are not
in meson-fallback downloads, such as when download pcre 8.37 from
https://ftp.pcre.org (the file is present, I verified, and I could also
download it on the remote build machine with curl, after connecting
through SSH). The error was:

> Looking for a fallback subproject for the dependency libpcre because:
> Use of fallback dependencies is forced.
> Downloading libpcre source from https://ftp.pcre.org/pub/pcre/pcre-8.37.tar.bz2
> <urlopen error [SSL: CERTIFICATE_VERIFY_FAILED] certificate verify failed: unable to get local issuer certificate (_ssl.c:1129)>
> A fallback URL could be specified using source_fallback_url key in the wrap file
> could not get https://ftp.pcre.org/pub/pcre/pcre-8.37.tar.bz2 is the internet available?
> Subproject  libpcre is buildable: NO (disabling)
> Dependency libpcre from subproject libpcre found: NO (subproject failed to configure)

This is probably linked to the recent fiasco for Let's Encrypt root
certificates not updated on time on various systems (this web server
does use Let's Encrypt after verifying).

After some web search, I found this command to update the certs used by
Python. It seemed to work fine when re-running the configuration through
SSH.