.circleci: use gimp-mac-bundler from my namespace.
We were using a gtk-mac-bundler from samm-git which was itself a fork of
a repository named the same from jralls on Github (itself a fork of some
other project). Unfortunately samm-git being absent lately, we don't
have access to his repository, so I fork it once more as
gimp-mac-bundler in my namespace.

Note that there is only a single commit over jralls' version (which
seems maintained) so ideally we should fist check if it's really still
needed. And if so, we should contribute it upstream and get rid of our
fork if it goes in (which I'm planning to act on).

For the time being, for sake of testing, I just use my fork, which I
just rebased over jralls' version as we have run errors with Python
3.9.7 from the CI server, which looks like they might be fixed by commit
58d951c on jralls' repo.

Errors were:

Creating bundle
Traceback (most recent call last):
  File "/Users/distiller/.local/bin/gtk-mac-bundler", line 5, in <module>
    import bundler.main
  File "/Users/distiller/Source/gtk-mac-bundler/bundler/main.py", line 3, in <module>
    from .project import *
  File "/Users/distiller/Source/gtk-mac-bundler/bundler/project.py", line 10, in <module>
    from plistlib import Plist
ImportError: cannot import name 'Plist' from 'plistlib' (/Library/Frameworks/Python.framework/Versions/3.9/lib/python3.9/plistlib.py)