gtk-osx: Move jhbuild modulesets into this project
By making this change, now all builds can be run locally, and there is
no need to go to an external project to make configuration changes.

This massively simplifies the job of managing releases and dependencies.

Also, as the files are accessed locally, there is not a problem with
hardcoded links and swapping around urls just to
test a minor change to see if it works.