gtk-osx: gtk3 fix a number of crashes
Continuing set of fixes related to problems with the font decorator
when adding a text box into the canvas and selecting fonts.

The quartz code in gtk 3 suffers from a number of reference
counting problems (or so it appears). This set of patches increases the
robustness of the pointer casting and falls back when certain pointers
are not set.

- Enhance gtk3 patch to zero out unused field to ensure it isn't used.
- NSWindows and NSViews appear to be getting lost track of.
- Document state of gtk patches upstream and if they have been merged

Fixes apply to: https://gitlab.gnome.org/GNOME/gimp/-/issues/7608