gimp: pygimp.interp correctly built
`pygimp.interp` is now correctly build in the gimp build process so no
longer needs to be patched in the packaging process.

The gimp build picks up the `$PYTHON` env variable (in autotools, to be
determined if this also happens in meson). This is then put into the
`pygimp.interp` file.