gtk-osx: Fix 2731
This patch correctly idenifies when a drag and drop failed so that the
correct messages are sent and the DND is cleaned up, including
removing a stray window that obscures the menu bar.

This does not fix the fact that the window is in the wrong place when it
is present.

The overall docking system still has some curious reliability issues,
these seem to now be recoverable. I've been able to get the
floating window to always disappear.