gtk-osx: Fix faulty AVIF support
Due to a complex chain of events, adding AVIF did not work.

The solution in the end involved:

- Forcing aom to compile its dylib in such a way that instead of putting
  in @rpath, it would put in the full install directory for its name.
- This involved calling `install_name_tool -id` with the correct path on
  the correct dylib
- This is required because the CMake instructions for putting the
  install path in executables did not work in any way
- And all this was required because the
  `gdk-pixbuf-query-loaders --update-cache` function would not
  work without a fully qualified path to the `libaom.3.2.0.dylib` which
  was not being installed in the `libheif` build process for the libheif
  loader due to it being `@rpath` in the actual dylib.

Additionally, updates info.plist for the new (and existing)file types:

- HEIC
- AVIF

Inspiration taken from `Preview.app` and `Krita.app` as well as
existing entries.

However, it isn't entirely clear how to set up the document.