macos: Fix bug 7690
This provides the foundation to implementing the slow drawing patch on
MacOS. This creates a new API function in `GtkLabel` that allows the
text in a label to be set without resizing the label.
This is critical because otherwise the label triggers a resize of the
entire drawing canvas (even when nothing has changed), which is
very slow.

*Note* This is an ugly monkey patch which is only possible because on
MacOS we are forced to bundle all our dependencies.
I was not able to find a way to contain the resizes within the GtkLabel
(I tried putting it in a `GtkScrollableWindow` and anything else
I could find. But nothing worked.

I also filed this bug: https://gitlab.gnome.org/GNOME/gtk/-/issues/4691
though I was not able to create a demo
program due to them not running correctly after building in JHBuild
(the unmodified examples).