circleci: Split build into jobs; auto cache break
This breaks the circleci build into multiple jobs which communicate
via cache. This means the 1 hour limit is no longer applicable.

However, the cost is that setup and cache restore must happen
on each build, which takes 5 minutes per job (4 * 5 min = 20 min).

This change also allows changes to the modulesets or patches to
auto break the caching scheme.

The algorithm is the following:

- On changes to the modulesets files or the `.circleci/config.yml`
  file, the caches are iterated, meaning
  they can be recached with new layered results ontop (save the
  cache again, but don't wipe it out).

- On changes to the patch files, the whole cache is wiped clean
  because jhbuild is unable to detect changes to the patches and
  will fail to rebuild the specific item without deleting the cached
  source. In future this could potentially be fixed by caching
  the source separately from the inst directory and only
  breaking the source cache.

circleci: disable gimp-help

Does not build

circleci: remove unsed packages