gtk-osx: jpeg-xl support
Adds support for jpeg-xl format.

Notes:

Locked `main` branch of `libjxl` to specific commit
so that additional commits do not break the build unexpectedly.
Should be locked to the tag of a release when 0.7.0 is released.

Build must stay git-based because it will not build without
the thirdparty submodules being available, even if they are not
properly installed (specifically, `libbrotli` is required to be there
and is built, but is not installed, so it also has to be built and
installed separately).